+++ packages/file_picker_android/example/example.dart
@@ -1,0 +1,7 @@
+// An example demonstrating the usage of android_file_picker.
+import 'package:android_file_picker/android_file_picker.dart';
+
+void main() {
+  FilePickerAndroid.registerWith();
+  print('Registered android_file_picker implementation.');
+}