+++ packages/file_picker_android/lib/android_file_picker.dart
@@ -1,3 +1,6 @@
+/// The Android implementation of the `file_picker` plugin.
+library;
+
 export 'src/android_platform_file.dart';
 export 'src/android_saf_handle.dart';
 export 'src/file_picker_android.dart';