+++ packages/file_picker_darwin/lib/file_picker_darwin.dart
@@ -1,2 +1,5 @@
+/// The Darwin (iOS and macOS) implementation of the `file_picker` plugin.
+library;
+
 export 'src/darwin_platform_file.dart';
 export 'src/file_picker_darwin.dart';