+++ packages/file_picker_linux/example/example.dart
@@ -1,0 +1,7 @@
+// An example demonstrating the usage of file_picker_linux.
+import 'package:file_picker_linux/file_picker_linux.dart';
+
+void main() {
+  FilePickerLinux.registerWith();
+  print('Registered file_picker_linux implementation.');
+}