+++ packages/file_picker_linux/lib/file_picker_linux.dart
@@ -1,3 +1,6 @@
+/// The Linux implementation of the `file_picker` plugin.
+library;
+
 export 'src/file_picker_linux.dart';
 export 'src/file_picker_linux_options.dart';
 export 'src/filters.dart';