+++ packages/file_picker_linux/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.1
 
 - Relax `dbus` to `^0.7.13` so dependents can resolve `xml` 7. [#2130](https://github.com/miguelpruivo/flutter_file_picker/issues/2130)
+- Improved package description, added example, and added missing API documentation.
 
 ## 1.0.0